requirements.txt
smolagents
requests
python-dotenv
